+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 12.12.0
+
+- Add `BOS_DEFAULT_LND_PATH` to set the default path to LND data directory
+- Adjust support for Umbrel LND path detection
+
 ## 12.11.2
 
 - `telegram`: Add inbound and outbound fee rates to /liquidity message